refactor: authorization functional tests
