+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/integration/AuthorizationFunctionalTest.java
@@ -16,21 +16,16 @@
 package io.confluent.ksql.rest.integration;
 
 import io.confluent.common.utils.IntegrationTest;
-import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.confluent.ksql.integration.IntegrationTestHarness;
-import io.confluent.ksql.rest.entity.CommandStatus;
-import io.confluent.ksql.rest.entity.CommandStatusEntity;
 import io.confluent.ksql.rest.entity.KafkaTopicInfo;
 import io.confluent.ksql.rest.entity.KafkaTopicsList;
 import io.confluent.ksql.rest.entity.KsqlEntity;
 import io.confluent.ksql.rest.server.TestKsqlRestApp;
 import io.confluent.ksql.rest.server.security.KsqlAuthorizationProvider;
-import io.confluent.ksql.rest.server.security.KsqlSecurityExtension;
 import io.confluent.ksql.services.ServiceContext;
 import io.confluent.ksql.test.util.secure.Credentials;
 import io.confluent.ksql.util.KsqlConfig;
-import io.confluent.ksql.util.KsqlException;
-import org.apache.kafka.streams.KafkaClientSupplier;
+import org.apache.kafka.common.errors.AuthorizationException;
 import org.glassfish.hk2.api.Factory;
 import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.process.internal.RequestScoped;
@@ -42,22 +37,26 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.RuleChain;
+import org.junit.runner.RunWith;
+import org.mockito.ArgumentMatcher;
 import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
 
-import javax.ws.rs.core.Configurable;
 import java.security.Principal;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.function.Supplier;
 
 import static io.confluent.ksql.test.util.EmbeddedSingleNodeKafkaCluster.JAAS_KAFKA_PROPS_NAME;
 import static io.confluent.ksql.test.util.EmbeddedSingleNodeKafkaCluster.VALID_USER1;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
 
 @Category({IntegrationTest.class})
+@RunWith(MockitoJUnitRunner.class)
 public class AuthorizationFunctionalTest {
   private static final IntegrationTestHarness TEST_HARNESS = IntegrationTestHarness.build();
 
@@ -96,8 +95,6 @@ public class AuthorizationFunctionalTest {
       })
       .build();
 
-  private static Set<String> allowedUsers = new HashSet<>();
-
   @Mock
   private KsqlAuthorizationProvider authorizationProvider;
 
@@ -114,7 +111,6 @@ public class AuthorizationFunctionalTest {
 
   @Before
   public void setUp() {
-    allowedUsers.clear();
     MockKsqlSecurityExtension.setAuthorizationProvider(authorizationProvider);
   }
 
@@ -127,6 +123,9 @@ public class AuthorizationFunctionalTest {
 
   @Test
   public void shouldDenyAccess() {
+    // Given:
+    denyAccess(USER1, "POST", "/ksql");
+
     // Then:
     expectedException.expect(AssertionError.class);
     expectedException.expectMessage(
@@ -140,7 +139,7 @@ public class AuthorizationFunctionalTest {
   @Test
   public void shouldAllowAccess() {
     // Given:
-    givenAuthorizedUser(USER1);
+    allowAccess(USER1, "POST", "/ksql");
 
     // When:
     final List<KsqlEntity> results = makeKsqlRequest(USER1, "SHOW TOPICS;");
@@ -151,60 +150,31 @@ public class AuthorizationFunctionalTest {
     assertThat(topics.get(0).getName(), is(TOPIC_1));
   }
 
-  private void givenAuthorizedUser(final Credentials user) {
-    allowedUsers.add(user.username);
+  private void allowAccess(final Credentials user, final String method, final String path) {
+    doNothing().when(authorizationProvider)
+        .checkEndpointAccess(argThat(new PrincipalMatcher(user)), eq(method), eq(path));
+  }
+
+  private void denyAccess(final Credentials user, final String method, final String path) {
+    doThrow(new AuthorizationException(String.format("Access denied to User:%s", user.username)))
+        .when(authorizationProvider)
+        .checkEndpointAccess(argThat(new PrincipalMatcher(user)), eq(method), eq(path));
   }
 
   private List<KsqlEntity> makeKsqlRequest(final Credentials credentials, final String sql) {
     return RestIntegrationTestUtil.makeKsqlRequest(REST_APP, sql, Optional.of(credentials));
   }
 
-  /*
-   * Mock the Security extension and authorization provider for all tests
-   */
-  public static class MockKsqlSecurityExtension implements KsqlSecurityExtension {
-    public static boolean initialized = false;
-    public static boolean closed = false;
-    public static KsqlAuthorizationProvider provider;
+  private static class PrincipalMatcher implements ArgumentMatcher<Principal> {
+    private final Credentials user;
 
-    public static void setAuthorizationProvider(final KsqlAuthorizationProvider provider) {
-      MockKsqlSecurityExtension.provider = provider;
+    PrincipalMatcher(final Credentials user) {
+      this.user = user;
     }
 
     @Override
-    public void initialize(KsqlConfig ksqlConfig) {
-      MockKsqlSecurityExtension.initialized = true;
-    }
-
-    @Override
-    public Optional<KsqlAuthorizationProvider> getAuthorizationProvider() {
-      return Optional.of((user, method, path) -> {
-            if (!allowedUsers.contains(user.getName())) {
-              throw new KsqlException(String.format("Access denied to User:%s", user));
-            }
-          }
-      );
-    }
-
-    @Override
-    public void register(Configurable<?> configurable) {
-
-    }
-
-    @Override
-    public KafkaClientSupplier getKafkaClientSupplier(Principal principal) throws KsqlException {
-      return null;
-    }
-
-    @Override
-    public Supplier<SchemaRegistryClient> getSchemaRegistryClientSupplier(Principal principal) throws KsqlException {
-      return null;
-    }
-
-    @Override
-    public void close() {
-      MockKsqlSecurityExtension.closed = true;
+    public boolean matches(final Principal principal) {
+      return this.user.username.equals(principal.getName());
     }
   }
-
 }
